+++ pkg/operatorV2/handle.go
@@ -20,7 +20,32 @@
 
 package operator
 
-import "context"
+import (
+	"context"
+
+	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
+)
+
+func WithCondition(conditions *api.ConditionList, condition api.ConditionType, changed bool, err error) (bool, error) {
+	if changed || err != nil {
+		// Condition should be false
+		if conditions.Update(condition, false, "Not ready", "Not ready") {
+			changed = true
+		}
+	} else {
+		if conditions.Update(condition, true, "Ready", "Ready") {
+			changed = true
+		}
+	}
+
+	if err == nil || IsStop(err) {
+		if changed {
+			err = Reconcile("Condition changed")
+		}
+	}
+
+	return changed, err
+}
 
 type HandleP0Func func(ctx context.Context) (bool, error)
 
@@ -66,6 +91,15 @@ func HandleP1[P1 interface{}](ctx context.Context, p1 P1, handler ...HandleP1Fun
 	return isChanged, nil
 }
 
+func HandleP1WithStop[P1 interface{}](ctx context.Context, p1 P1, handler ...HandleP1Func[P1]) (bool, error) {
+	changed, err := HandleP1[P1](ctx, p1, handler...)
+	if IsStop(err) {
+		return changed, nil
+	}
+
+	return changed, err
+}
+
 func HandleP2[P1, P2 interface{}](ctx context.Context, p1 P1, p2 P2, handler ...HandleP2Func[P1, P2]) (bool, error) {
 	isChanged := false
 	for _, h := range handler {
@@ -80,6 +114,15 @@ func HandleP2[P1, P2 interface{}](ctx context.Context, p1 P1, p2 P2, handler ...
 	}
 
 	return isChanged, nil
+}
+
+func HandleP2WithStop[P1, P2 interface{}](ctx context.Context, p1 P1, p2 P2, handler ...HandleP2Func[P1, P2]) (bool, error) {
+	changed, err := HandleP2[P1, P2](ctx, p1, p2, handler...)
+	if IsStop(err) {
+		return changed, nil
+	}
+
+	return changed, err
 }
 
 func HandleP3[P1, P2, P3 interface{}](ctx context.Context, p1 P1, p2 P2, p3 P3, handler ...HandleP3Func[P1, P2, P3]) (bool, error) {
@@ -98,6 +141,20 @@ func HandleP3[P1, P2, P3 interface{}](ctx context.Context, p1 P1, p2 P2, p3 P3, 
 	return isChanged, nil
 }
 
+func HandleP3WithStop[P1, P2, P3 interface{}](ctx context.Context, p1 P1, p2 P2, p3 P3, handler ...HandleP3Func[P1, P2, P3]) (bool, error) {
+	changed, err := HandleP3[P1, P2, P3](ctx, p1, p2, p3, handler...)
+	if IsStop(err) {
+		return changed, nil
+	}
+
+	return changed, err
+}
+
+func HandleP3WithCondition[P1, P2, P3 interface{}](ctx context.Context, conditions *api.ConditionList, condition api.ConditionType, p1 P1, p2 P2, p3 P3, handler ...HandleP3Func[P1, P2, P3]) (bool, error) {
+	changed, err := HandleP3[P1, P2, P3](ctx, p1, p2, p3, handler...)
+	return WithCondition(conditions, condition, changed, err)
+}
+
 func HandleP4[P1, P2, P3, P4 interface{}](ctx context.Context, p1 P1, p2 P2, p3 P3, p4 P4, handler ...HandleP4Func[P1, P2, P3, P4]) (bool, error) {
 	isChanged := false
 	for _, h := range handler {
@@ -114,6 +171,20 @@ func HandleP4[P1, P2, P3, P4 interface{}](ctx context.Context, p1 P1, p2 P2, p3 
 	return isChanged, nil
 }
 
+func HandleP4WithStop[P1, P2, P3, P4 interface{}](ctx context.Context, p1 P1, p2 P2, p3 P3, p4 P4, handler ...HandleP4Func[P1, P2, P3, P4]) (bool, error) {
+	changed, err := HandleP4[P1, P2, P3, P4](ctx, p1, p2, p3, p4, handler...)
+	if IsStop(err) {
+		return changed, nil
+	}
+
+	return changed, err
+}
+
+func HandleP4WithCondition[P1, P2, P3, P4 interface{}](ctx context.Context, conditions *api.ConditionList, condition api.ConditionType, p1 P1, p2 P2, p3 P3, p4 P4, handler ...HandleP4Func[P1, P2, P3, P4]) (bool, error) {
+	changed, err := HandleP4[P1, P2, P3, P4](ctx, p1, p2, p3, p4, handler...)
+	return WithCondition(conditions, condition, changed, err)
+}
+
 func HandleP9[P1, P2, P3, P4, P5, P6, P7, P8, P9 interface{}](ctx context.Context, p1 P1, p2 P2, p3 P3, p4 P4, p5 P5, p6 P6, p7 P7, p8 P8, p9 P9, handler ...HandleP9Func[P1, P2, P3, P4, P5, P6, P7, P8, P9]) (bool, error) {
 	isChanged := false
 	for _, h := range handler {
@@ -128,4 +199,13 @@ func HandleP9[P1, P2, P3, P4, P5, P6, P7, P8, P9 interface{}](ctx context.Contex
 	}
 
 	return isChanged, nil
+}
+
+func HandleP9WithStop[P1, P2, P3, P4, P5, P6, P7, P8, P9 interface{}](ctx context.Context, p1 P1, p2 P2, p3 P3, p4 P4, p5 P5, p6 P6, p7 P7, p8 P8, p9 P9, handler ...HandleP9Func[P1, P2, P3, P4, P5, P6, P7, P8, P9]) (bool, error) {
+	changed, err := HandleP9[P1, P2, P3, P4, P5, P6, P7, P8, P9](ctx, p1, p2, p3, p4, p5, p6, p7, p8, p9, handler...)
+	if IsStop(err) {
+		return changed, nil
+	}
+
+	return changed, err
 }